backend/requirements.txt
Django==5.1.4
psycopg2-binary==2.9.10
Pillow==11.0.0
celery==5.3.6
redis==5.0.2
djangorestframework==3.15.2
boto3==1.26.59
django-storages==1.14.4
